# specs/001-npd-v1-core/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Novus Project Database API
  description: |
    Internal REST API for the Novus Project Database (NPD) v1.
    Provides project management, document storage, search, and RAG-assisted import capabilities.
  version: 1.0.0
  contact:
    name: NPD Development Team
servers:
  - url: /api/v1
    description: API base path
...
paths:
...
  /auth/callback:
    get:
      tags: [auth]
      summary: Azure AD OAuth callback
      description: Handles Azure AD redirect after authentication
      operationId: authCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect to application with session cookie
  /auth/logout:
    post:
      tags: [auth]
      summary: Log out current user
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully logged out
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user info
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # Projects
  # ============================================================
...
  /organizations:
    get:
      tags: [organizations]
      summary: List organizations
      operationId: listOrganizations
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: search
          in: query
          description: Search by name or alias
          schema:
            type: string
      responses:
        '200':
          description: Organization list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [organizations]
      summary: Create an organization
      operationId: createOrganization
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{organization_id}:
    parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [organizations]
      summary: Get organization details
      operationId: getOrganization
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Organization details with project count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [organizations]
      summary: Update an organization
      operationId: updateOrganization
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdate'
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # Contacts
  # ============================================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    UserRole:
      type: string
      enum: [user, admin]

    # --- Core Entities ---
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
      required: [id, email, display_name, role]
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        aliases:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name]
    OrganizationCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        aliases:
          type: array
          items:
            type: string
      required: [name]
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        aliases:
          type: array
          items:
            type: string
    OrganizationDetail:
      allOf:
        - $ref: '#/components/schemas/Organization'
        - type: object
          properties:
            project_count:
              type: integer
            contacts:
              type: array
              items:
                $ref: '#/components/schemas/Contact'
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        organization_id:
          type: string
          format: uuid
        organization:
          $ref: '#/components/schemas/Organization'
        role_title:
          type: string
        phone:
          type: string
        notes:
          type: string
        monday_url:
          type: string
          format: uri
      required: [id, name, email, organization_id]
...
    OrganizationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
      required: [items, total, page, page_size]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          default: validation_error
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required: [error, message]